pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "logfsm"
version = "0.1.0"
description = "Log-to-FSM pipeline with rule suggestion and FSM inference"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "PyYAML>=6.0",
]

[project.scripts]
logfsm = "logfsm.cli:main"
